refactor - move longer styles to their own file
