+++ client/src/components/CalendarView/CalendarWeek.tsx
@@ -1,49 +1,12 @@
-import styled from 'styled-components';
 import CalendarDay from 'components/CalendarView/CalendarDay';
 import {Week} from 'types/Week';
+import {CalendarWeekStyles} from 'styles/CalendarWeekStyles';
 import leftArrowIcon from 'images/calendar/leftArrow.svg'
 import rightArrowIcon from 'images/calendar/rightArrow.svg'
 
 interface CalendarWeekProps {
     week: Week
 }
-
-const CalendarWeekStyles = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin: 20px;
-
-  .calendar-header {
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 50px;
-  }
-
-  .calendar-header-month {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 10px;
-  }
-
-  img {
-    width: 30px;
-    height: 30px;
-    background-color: var(--white);
-    border-radius: 9999px;
-    cursor: pointer;
-  }
-
-  h1 {
-    padding-bottom: 5px;
-  }
-
-  .lesson-grid {
-    display: flex;
-    flex-direction: row;
-    gap: 20px;
-  }
-`
 
 const CalendarWeek = ({week}: CalendarWeekProps) => {
     return (
@@ -55,14 +18,12 @@ const CalendarWeek = ({week}: CalendarWeekProps) => {
                         <img src={leftArrowIcon} alt={`back-arrow`}/>
                         <img src={rightArrowIcon} alt={`forward-arrow`}/>
                     </div>
-
                     <div>Week {week.dateRange}</div>
                 </div>
                 <div>
                     <button>Add Lesson</button>
                 </div>
             </div>
-
             <div className="lesson-grid">
                 {week && week.days.map((day) => (
                     <CalendarDay key={day.day} day={day}/>
