+++ client/src/components/Sidebar/UserLogout.tsx
@@ -1,40 +1,6 @@
-import styled from 'styled-components';
+import {UserLogoutStyles} from 'styles/UserLogoutStyles';
 import avatar from 'images/sidebar/user.svg'
 import logoutIcon from 'images/sidebar/logout.svg'
-
-
-const UserLogoutStyles = styled.div`
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: row;
-  gap: 10px;
-  align-items: flex-end;
-  justify-content: space-between;
-  width: 100%;
-  padding: 30px 10px;
-
-  img {
-    margin: auto;
-  }
-
-  .img-container {
-    display: flex;
-    width: 30px;
-    height: 30px;
-    padding: 5px;
-    border-radius: 9999px;
-    background-color: var(--background-color)
-  }
-
-  .logout {
-    background-color: var(--white);
-    width: 20px;
-    height: 20px;
-  }
-
-
-`
-
 
 const UserLogout = () => {
     return (
@@ -49,7 +15,7 @@ const UserLogout = () => {
             <div className="img-container logout">
                 <img src={logoutIcon} alt={`logout-icon`}/>
             </div>
-
+            s
         </UserLogoutStyles>
     )
 }
